pyproject.toml
[tool.poetry]
name = "deep-factors"
version = "0.1.0"
description = "master"
authors = ["Pablo Calvache"]
license = "free"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.13"
tensorflow-macos = "2.14.0"
numpy = "1.24.3"
jupyter = "^1.0.0"
tensorflow-metal = "^1.1.0"
pandas = "2.0.3"
scikit-learn = "^1.3.1"
matplotlib = "^3.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
